// packages/shared/src/constants.d.ts
export declare const BUSINESS_HOURS: {
    open: number;
    close: number;
};
export declare const SLOT_DURATION = 60;
export declare const PEAK_START = 17;
export declare const PEAK_END = 24;
export declare const WEEKEND_DAYS: number[];
export declare const BUSINESS_LOCATION: {
    name: string;
    name_ar: string;
    address: string;
    address_ar: string;
    lat: number;
    lng: number;
    maps_link: string;
    phone: string;
    whatsapp: string;
};
export declare const SEGMENT_THRESHOLDS: {
    new: {
        min: number;
        max: number;
    };
    occasional: {
        min: number;
        max: number;
    };
    regular: {
        min: number;
        max: number;
    };
    vip: {
        min: number;
        max: number;
    };
};
export declare const CURRENCY = "JOD";
export declare const CURRENCY_AR = "\u062F\u064A\u0646\u0627\u0631";